ip_adapter/disentangling-vae-master2/requirements.txt
absl-py==1.4.0
alembic==1.7.7
appdirs==1.4.4
asn1crypto==0.24.0
attrs==22.2.0
audioread==2.1.9
autopage==0.5.1
bottle==0.12.13
cachetools==4.2.4
catfish==1.4.4
certifi==2018.1.18
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==6.7
cliff==3.10.1
cloudpickle==2.2.1
cmaes==0.9.0
cmd2==2.4.3
colorama==0.3.9
colorlog==6.7.0
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defer==1.0.6
#disentanglement-lib @ file:///home/santiagojn/lib_disentanglement
distro-info===0.18ubuntu0.18.04.1
docker==2.5.1
docker-pycreds==0.2.1
et-xmlfile==1.1.0
future==0.18.2
gin-config==0.5.0
Glances==2.11.1
google-auth==2.16.3
google-auth-oauthlib==0.4.6
greenlet==2.0.2
grpcio==1.48.2
httplib2==0.9.2
idna==2.6
imageio==2.15.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
influxdb==4.1.1
iotop==0.6
ipython==5.5.0
ipython_genutils==0.2.0
joblib==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
librosa==0.9.1
lightdm-gtk-greeter-settings==1.2.2
llvmlite==0.36.0
Mako==1.1.6
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
meld==3.18.0
menulibre==2.2.0
mugshot==0.4.0
natsort==8.2.0
netifaces==0.10.4
networkx==2.5.1
numba==0.53.1
numpy==1.19.5
oauth==1.0.1
oauthlib==3.2.2
olefile==0.45.1
onboard==1.4.1
opencv-python==4.5.5.64
OpenEXR==1.3.9
openpyxl==3.1.2
optuna==3.0.6
packaging==21.3
pandas==1.1.5
pbr==5.11.1
pexpect==4.2.1
pickle5==0.0.12
pickleshare==0.7.4
Pillow==8.4.0
pluggy==0.6.0
ply==3.11
pooch==1.6.0
prettytable==2.5.0
prompt-toolkit==1.0.15
protobuf==3.19.6
psutil==5.4.2
py==1.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
pycryptodomex==3.4.7
pycups==1.9.73
pycurl==7.43.0.1
Pygments==2.2.0
PyGObject==3.26.1
pynvml==11.4.1
pyparsing==2.2.0
pyperclip==1.8.2
PySimpleSOAP==1.16.2
pysmi==0.2.2
pysnmp==4.4.3
pystache==0.5.4
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.2
python-debian==0.1.32
python-debianbts==2.7.2
python-magic==0.4.16
pytz==2018.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.1.5
resampy==0.2.2
rsa==4.9
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
screen-resolution-extra==0.0.0
seaborn==0.11.2
SecretStorage==2.3.1
sgt-launcher==0.2.4
simplegeneric==0.8.1
simplejson==3.13.2
six==1.16.0
SoundFile==0.10.3.post1
SQLAlchemy==1.4.49
ssh-import-id==5.7
stevedore==3.5.2
systemd-python==234
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-hub==0.15.0
tensorflow-probability==0.7.0
testresources==2.0.1
threadpoolctl==3.1.0
tifffile==2020.9.3
torch==1.5.0+cu101
torchvision==0.6.0+cu101
tqdm==4.64.0
traitlets==4.3.2
typing_extensions==4.1.1
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
wadllib==1.3.2
wcwidth==0.1.7
websocket-client==0.44.0
Werkzeug==2.0.3
xcffib==0.5.1
xkit==0.0.0
zipp==3.6.0
zope.interface==4.3.2
